Created View namespace, modified references for the mainpage and dashboard views.
