+++ NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay/MainPage.xaml.cs
@@ -1,10 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Xamarin.Forms;
+using NeverSkipLegDay.Views;
 
 namespace NeverSkipLegDay
 {
