+++ NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay.csproj.user
@@ -2,7 +2,7 @@
 <Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup />
   <ItemGroup>
-    <Compile Update="DashboardMasterMenuItem.cs">
+    <Compile Update="Views\DashboardMasterMenuItem.cs">
       <SubType>Code</SubType>
     </Compile>
   </ItemGroup>
