+++ NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay/Views/Dashboard.xaml
@@ -4,8 +4,8 @@
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d"
-             x:Class="NeverSkipLegDay.Dashboard"
-             xmlns:pages="clr-namespace:NeverSkipLegDay">
+             x:Class="NeverSkipLegDay.Views.Dashboard"
+             xmlns:pages="clr-namespace:NeverSkipLegDay.Views">
   <MasterDetailPage.Master>
     <pages:DashboardMaster x:Name="MasterPage" />
   </MasterDetailPage.Master>
