+++ NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay/Views/DashboardDetail.xaml
@@ -4,7 +4,7 @@
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d"
-             x:Class="NeverSkipLegDay.DashboardDetail" >
+             x:Class="NeverSkipLegDay.Views.DashboardDetail" >
   <NavigationPage.TitleView>
     <Label
       FontAttributes="Bold"
