+++ NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay/Views/DashboardMaster.xaml
@@ -4,7 +4,7 @@
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d"
-             x:Class="NeverSkipLegDay.DashboardMaster"
+             x:Class="NeverSkipLegDay.Views.DashboardMaster"
              Title="Master">
   <StackLayout BackgroundColor="{StaticResource CardBackgroundColor}">
     <ListView x:Name="MenuItemsListView"
