+++ NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay/Views/Workouts/AddEditWorkout.xaml
@@ -2,16 +2,16 @@
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"     
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              BackgroundColor="{StaticResource WorkoutsBackgroundColor}"
              mc:Ignorable="d"
-             x:Class="NeverSkipLegDay.Workouts.EditExercise">
+             x:Class="NeverSkipLegDay.Views.AddEditWorkout">
   <NavigationPage.TitleView>
     <Label 
       FontAttributes="Bold"
       TextColor="{StaticResource BarTextColor}"
       FontSize="20" 
-      Text="E X E R C I S E S"
+      Text="W O R K O U T S"
       HorizontalOptions="Center"
       Margin="0,0,70,0"/>
   </NavigationPage.TitleView>
@@ -25,23 +25,16 @@
         FontSize="20"
         Margin="10"
         TextColor="{StaticResource CardTextColor}"/>
-      <StackLayout
-        Orientation="Horizontal">
-        <StackLayout
-          HorizontalOptions="FillAndExpand">
-          <Entry 
-            Text="{Binding Name}"
-            Placeholder="e.g. Bench Press" 
-            TextColor="{StaticResource CardTextColor}"
-            PlaceholderColor="{StaticResource PlaceholderColor}"
-            Margin="10,0,10,10" 
-            IsTextPredictionEnabled="True"  
-            Keyboard="Chat" 
-            IsSpellCheckEnabled="True"/>
-        </StackLayout>
-      </StackLayout>
-      <Button
-        BindingContext="{Binding ID}"
+      <Entry 
+        Text="{Binding Name}"
+        TextColor="{StaticResource CardTextColor}"
+        PlaceholderColor="{StaticResource PlaceholderColor}"
+        Placeholder="e.g. Push Day" 
+        Margin="10,0,10,10" 
+        IsTextPredictionEnabled="True"  
+        Keyboard="Chat" 
+        IsSpellCheckEnabled="True" />
+      <Button 
         Text="Save"
         HorizontalOptions="End"
         Margin="0,0,10,10"
