+++ NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay/Views/Workouts/EditExercise.xaml
@@ -2,16 +2,16 @@
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"     
              BackgroundColor="{StaticResource WorkoutsBackgroundColor}"
              mc:Ignorable="d"
-             x:Class="NeverSkipLegDay.Workouts.AddEditWorkout">
+             x:Class="NeverSkipLegDay.Views.EditExercise">
   <NavigationPage.TitleView>
     <Label 
       FontAttributes="Bold"
       TextColor="{StaticResource BarTextColor}"
       FontSize="20" 
-      Text="W O R K O U T S"
+      Text="E X E R C I S E S"
       HorizontalOptions="Center"
       Margin="0,0,70,0"/>
   </NavigationPage.TitleView>
@@ -25,16 +25,23 @@
         FontSize="20"
         Margin="10"
         TextColor="{StaticResource CardTextColor}"/>
-      <Entry 
-        Text="{Binding Name}"
-        TextColor="{StaticResource CardTextColor}"
-        PlaceholderColor="{StaticResource PlaceholderColor}"
-        Placeholder="e.g. Push Day" 
-        Margin="10,0,10,10" 
-        IsTextPredictionEnabled="True"  
-        Keyboard="Chat" 
-        IsSpellCheckEnabled="True" />
-      <Button 
+      <StackLayout
+        Orientation="Horizontal">
+        <StackLayout
+          HorizontalOptions="FillAndExpand">
+          <Entry 
+            Text="{Binding Name}"
+            Placeholder="e.g. Bench Press" 
+            TextColor="{StaticResource CardTextColor}"
+            PlaceholderColor="{StaticResource PlaceholderColor}"
+            Margin="10,0,10,10" 
+            IsTextPredictionEnabled="True"  
+            Keyboard="Chat" 
+            IsSpellCheckEnabled="True"/>
+        </StackLayout>
+      </StackLayout>
+      <Button
+        BindingContext="{Binding ID}"
         Text="Save"
         HorizontalOptions="End"
         Margin="0,0,10,10"
