+++ NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay/Views/Workouts/ExerciseProgress.xaml
@@ -5,7 +5,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              BackgroundColor="{StaticResource WorkoutsBackgroundColor}"
              mc:Ignorable="d"
-             x:Class="NeverSkipLegDay.Workouts.ExerciseProgress">
+             x:Class="NeverSkipLegDay.Views.ExerciseProgress">
   <NavigationPage.TitleView>
     <Label 
       FontAttributes="Bold"
