+++ NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay/Views/Workouts/Workout.xaml.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 
-namespace NeverSkipLegDay.Workouts
+namespace NeverSkipLegDay.Views
 {
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class Workout : ContentPage
